day 19 improvements
compute invariant signatures of each scanner and use them
to figure out which scanners overlap

this shaves off a factor of n from our main loop by not having to
hunt through the unaligned scanners, plus another factor of n by not
having to compare with all of the aligned scanners. the cost of
precomputing the overlaps is an o(n^2 b^2) loop (where b is the avg
number of beacons per scanner) but it's fast so we don't care.

inspired by rsc's ivy solution:
https://www.youtube.com/watch?v=o6GoqXcXsp0